kunrequirements.txt
discord.py==1.7.3
pypi-json==0.2.0
python-dotenv==0.19.2
datetime==4.3
asyncio==3.4.3
DiscordUtils==1.3.4
python-math==0.0.1
pandas==1.3.4
matplotlib==3.5.0